[Metal] refactor enum for graphics interface
- Move GraphicsAPI to own file
